requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
python-dotenv>=1.0.0

# AI APIs
google-generativeai>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
